--- site/components/headings/style.css
@@ -1,29 +1,0 @@
-.fontXL {
-  composes: fontXL from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
-}
-
-.fontL {
-  composes: fontL from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
-}
-
-.fontM {
-  composes: fontM from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
-}
-
-.fontM2 {
-  composes: fontM2 from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
-}
-
-.fontS2 {
-  composes: fontS2 from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
-}
-
-.fontS {
-  composes: fontS from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
-}
