+++ site/components/text/style.css
@@ -1,5 +1,35 @@
 
-.xSmallText {
+.fontXL {
+  composes: fontXL from "../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+}
+
+.fontL {
+  composes: fontL from "../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+}
+
+.fontM {
+  composes: fontM from "../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+}
+
+.fontM2 {
+  composes: fontM2 from "../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+}
+
+.fontS2 {
+  composes: fontS2 from "../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+}
+
+.fontS {
+  composes: fontS from "../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+}
+
+.fontXS {
   composes: fontXS from "../../css/typography/_fonts.css";
   composes: sansSerif from "../../css/typography/_fonts.css";
 }
